--- a/src/libraries/Common/src/Interop/Unix/System.Native/Interop.Socket.cs
+++ b/src/libraries/Common/src/Interop/Unix/System.Native/Interop.Socket.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Net;
-using System.Net.Internals;
 using System.Net.Sockets;
 using System.Runtime.InteropServices;
 
@@ -11,7 +10,11 @@ internal static partial class Interop
 {
     internal static partial class Sys
     {
+#if SYSTEM_NET_SOCKETS_DLL
         [LibraryImport(Libraries.SystemNative, EntryPoint = "SystemNative_Socket")]
         internal static unsafe partial Error Socket(AddressFamily addressFamily, SocketType socketType, ProtocolType protocolType, IntPtr* socket);
+#endif
+        [LibraryImport(Libraries.SystemNative, EntryPoint = "SystemNative_Socket")]
+        internal static unsafe partial Error Socket(int addressFamily, int socketType, int protocolType, IntPtr* socket);
     }
 }