--- a/src/libraries/Common/src/Interop/Unix/System.Native/Interop.Socket.cs
+++ b/src/libraries/Common/src/Interop/Unix/System.Native/Interop.Socket.cs
@@ -10,10 +10,6 @@ internal static partial class Interop
 {
     internal static partial class Sys
     {
-#if SYSTEM_NET_SOCKETS_DLL
-        [LibraryImport(Libraries.SystemNative, EntryPoint = "SystemNative_Socket")]
-        internal static unsafe partial Error Socket(AddressFamily addressFamily, SocketType socketType, ProtocolType protocolType, IntPtr* socket);
-#endif
         [LibraryImport(Libraries.SystemNative, EntryPoint = "SystemNative_Socket")]
         internal static unsafe partial Error Socket(int addressFamily, int socketType, int protocolType, IntPtr* socket);
     }