--- a/src/libraries/Common/src/Interop/Windows/WinSock/Interop.WSAConnect.cs
+++ b/src/libraries/Common/src/Interop/Windows/WinSock/Interop.WSAConnect.cs
@@ -10,13 +10,22 @@ internal static partial class Interop
     internal static partial class Winsock
     {
         [LibraryImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static partial SocketError WSAConnect(
+        private static partial SocketError WSAConnect(
             SafeSocketHandle socketHandle,
-            Span<byte> socketAddress,
+            ReadOnlySpan<byte> socketAddress,
             int socketAddressSize,
             IntPtr inBuffer,
             IntPtr outBuffer,
             IntPtr sQOS,
             IntPtr gQOS);
+
+        internal static SocketError WSAConnect(
+            SafeSocketHandle socketHandle,
+            ReadOnlySpan<byte> socketAddress,
+            IntPtr inBuffer,
+            IntPtr outBuffer,
+            IntPtr sQOS,
+            IntPtr gQOS) =>
+            WSAConnect(socketHandle, socketAddress, socketAddress.Length, inBuffer, outBuffer, sQOS, gQOS);
     }
 }