--- a/src/libraries/Common/src/Interop/Windows/WinSock/Interop.accept.cs
+++ b/src/libraries/Common/src/Interop/Windows/WinSock/Interop.accept.cs
@@ -12,7 +12,7 @@ internal static partial class Interop
         [LibraryImport(Interop.Libraries.Ws2_32, SetLastError = true)]
         internal static partial IntPtr accept(
             SafeSocketHandle socketHandle,
-            byte[] socketAddress,
+            Span<byte> socketAddress,
             ref int socketAddressSize);
     }
 }