--- a/src/libraries/Common/src/Interop/Windows/WinSock/Interop.sendto.cs
+++ b/src/libraries/Common/src/Interop/Windows/WinSock/Interop.sendto.cs
@@ -15,7 +15,7 @@ internal static partial class Interop
             byte* pinnedBuffer,
             int len,
             SocketFlags socketFlags,
-            byte[] socketAddress,
+            ReadOnlySpan<byte> socketAddress,
             int socketAddressSize);
     }
 }