--- a/src/libraries/Common/src/System/Net/IPEndPointExtensions.cs
+++ b/src/libraries/Common/src/System/Net/IPEndPointExtensions.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics;
 using System.Net;
 
 namespace System.Net.Sockets