… src/libraries/Common/src/System/Net/SocketAddressPal.Unix.cs
… src/libraries/Common/src/System/Net/SocketProtocolSupportPal.Unix.cs
… src/libraries/System.Net.NameResolution/tests/PalTests/System.Net.NameResolution.Pal.Tests.csproj
… src/libraries/System.Net.NetworkInformation/src/System.Net.NetworkInformation.csproj
… src/libraries/System.Net.NetworkInformation/src/System/Net/NetworkInformation/SystemNetworkInterface.cs
… src/libraries/System.Net.Ping/src/System.Net.Ping.csproj
… src/libraries/System.Net.Ping/src/System/Net/NetworkInformation/Ping.Windows.cs
… src/libraries/System.Net.Primitives/ref/System.Net.Primitives.cs
… src/libraries/System.Net.Primitives/tests/FunctionalTests/SocketAddressTest.cs
… src/libraries/System.Net.Quic/src/System.Net.Quic.csproj
… src/libraries/System.Net.Quic/src/System/Net/Quic/Internal/MsQuicExtensions.cs
… src/libraries/System.Net.Quic/src/System/Net/Quic/Internal/MsQuicHelpers.cs
… src/libraries/System.Net.Quic/src/System/Net/Quic/QuicConnection.cs
… src/libraries/System.Net.Quic/src/System/Net/Quic/QuicListener.cs
… src/libraries/System.Net.Sockets/ref/System.Net.Sockets.cs
… src/libraries/System.Net.Sockets/src/Resources/Strings.resx
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/Socket.Windows.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/Socket.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/SocketAsyncContext.Unix.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/SocketAsyncEventArgs.Unix.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/SocketAsyncEventArgs.Windows.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/SocketPal.Unix.cs
… src/libraries/System.Net.Sockets/src/System/Net/Sockets/SocketPal.Windows.cs
… src/libraries/System.Net.Sockets/tests/FunctionalTests/ReceiveFrom.cs
… src/native/libs/System.Native/entrypoints.c
… src/native/libs/System.Native/pal_networking.c
… src/native/libs/System.Native/pal_networking.h
… src/native/libs/System.Native/pal_networking_wasi.c