And uses new SocketAddress API. That spanifies some of the PAL code as prep for removal of Internal.SocketAddress. 

fixes  #86872 
contributes to  #87397
contributes to https://github.com/dotnet/runtime/issues/30797