@wfurt 2023-07-27
> I assume SendToAsync / ReceiveFromAsync are coming in a follow-up PR?

yes. That needs more work to clean up Internal.SocketAddress in socket code.