@wfurt 2023-08-01
perf check

|                                         Method |        Job |          Toolchain |         Mean |       Error |      StdDev |       Median |          Min |          Max | Ratio | RatioSD | Allocated | Alloc Ratio |
|----------------------------------------------- |----------- |------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|------:|--------:|----------:|------------:|
|                             ConnectAcceptAsync | Job-QAPYHK | \PR   \corerun.exe |     416.5 us |     4.92 us |     4.61 us |     415.9 us |     408.6 us |     426.5 us |  1.01 |    0.01 |    1370 B |        0.99 |
|                             ConnectAcceptAsync | Job-ZQMRZG |  \main\corerun.exe |     411.2 us |     3.12 us |     2.77 us |     410.7 us |     407.7 us |     416.6 us |  1.00 |    0.00 |    1386 B |        1.00 |
|                                                |            |                    |              |             |             |              |              |              |       |         |           |             |
|                 SendAsyncThenReceiveAsync_Task | Job-QAPYHK | \PR   \corerun.exe | 271,064.3 us | 4,562.00 us | 4,267.30 us | 269,604.6 us | 265,718.7 us | 277,834.9 us |  1.01 |    0.02 |     752 B |        1.00 |
|                 SendAsyncThenReceiveAsync_Task | Job-ZQMRZG |  \main\corerun.exe | 267,751.9 us | 1,734.53 us | 1,622.48 us | 267,574.0 us | 265,252.0 us | 270,458.9 us |  1.00 |    0.00 |     752 B |        1.00 |
|                                                |            |                    |              |             |             |              |              |              |       |         |           |             |
|                 ReceiveAsyncThenSendAsync_Task | Job-QAPYHK | \PR   \corerun.exe | 351,264.2 us | 1,238.80 us | 1,158.77 us | 351,484.9 us | 349,380.1 us | 353,069.9 us |  1.02 |    0.01 |    1280 B |        1.29 |
|                 ReceiveAsyncThenSendAsync_Task | Job-ZQMRZG |  \main\corerun.exe | 345,229.4 us | 1,554.32 us | 1,453.92 us | 345,121.6 us | 343,334.1 us | 348,427.2 us |  1.00 |    0.00 |     992 B |        1.00 |
|                                                |            |                    |              |             |             |              |              |              |       |         |           |             |
| SendAsyncThenReceiveAsync_SocketAsyncEventArgs | Job-QAPYHK | \PR   \corerun.exe | 262,947.6 us | 2,276.07 us | 2,129.04 us | 261,991.3 us | 260,505.5 us | 266,591.9 us |  0.96 |    0.01 |    1920 B |        1.00 |
| SendAsyncThenReceiveAsync_SocketAsyncEventArgs | Job-ZQMRZG |  \main\corerun.exe | 274,286.3 us | 2,102.03 us | 1,966.24 us | 274,853.0 us | 270,239.3 us | 277,223.5 us |  1.00 |    0.00 |    1920 B |        1.00 |
|                                                |            |                    |              |             |             |              |              |              |       |         |           |             |
| ReceiveAsyncThenSendAsync_SocketAsyncEventArgs | Job-QAPYHK | \PR   \corerun.exe | 331,678.8 us |   990.62 us |   926.63 us | 331,653.9 us | 329,933.3 us | 333,237.0 us |  1.00 |    0.01 |    3088 B |        1.33 |
| ReceiveAsyncThenSendAsync_SocketAsyncEventArgs | Job-ZQMRZG |  \main\corerun.exe | 332,023.6 us | 1,PR   1 us | 1,428.75 us | 332,203.7 us | 329,059.4 us | 334,365.6 us |  1.00 |    0.00 |    2320 B |        1.00 |